requirements.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
idna==3.7
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
gspread-formatting==1.2.0
idna==3.7
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
gspread-formatting==1.2.0
idna==3.7
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
gspread-formatting==1.2.0
idna==3.7
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
darkdetect==0.8.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
gspread-formatting==1.2.0
idna==3.7
oauthlib==3.2.2
packaging==24.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.2
customtkinter~=5.2.2
pip~=23.3
flet~=0.23.2
Jinja2~=3.1.4